app/requirements.txt
Django==4.1.2
Pillow==9.2.0
djangorestframework==3.14.0
psycopg2-binary==2.9.4